requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
email-validator==2.0.0
Werkzeug==2.3.7
Flask-Login==0.6.2
SQLAlchemy==2.0.21
waitress==2.1.2
python-dotenv==1.0.0
Flask-Limiter==3.5.0
Flask-Caching==2.0.2
python-magic==0.4.27
python-magic-bin==0.4.14
colorama==0.4.6